fix(reproduce): faithfully reproduce mm-composed figures (all panels, not one)
FigureRecord.from_dict re-keyed EVERY axes to grid_id(row,col); mm-composed
figures key panels "ax_mm_0..N" (no grid position) so they all collapsed to
"r0c0" -> a composed figure reproduced as a single panel at the wrong size
(e.g. 1138x898 vs 1874x1080, only the last panel surviving).

Fix:
- from_dict preserves non-grid keys verbatim (grid keys still normalize).
- New _reproducer/_mm_compose: rebuild composed figures by add_axes(bbox) per
  recorded panel, then replay. Dispatched from reproduce_from_record when all
  axes are ax_mm_*.
- Split over-limit _recorder/_core.py: Recorder engine -> _recorder/_recorder.py
  (re-exported), leaving the record dataclasses in _core.

Composed Fig1 now reproduces all 3 panels correctly (size 1871x1075 vs
1874x1080; residual is sub-pixel auto-crop rounding). No new test failures
(16 pre-existing TestPixelPerfect/TestMetadataRoundtrip only).
