+++ src/figrecipe/_recorder/__init__.py
@@ -2,7 +2,8 @@
 # -*- coding: utf-8 -*-
 """Recording infrastructure for figrecipe."""
 
-from ._core import AxesRecord, CallRecord, FigureRecord, Recorder
+from ._core import AxesRecord, CallRecord, FigureRecord
+from ._recorder import Recorder
 from ._utils import process_args
 
 __all__ = [
